refactor(mcp): modularize MCP server into separate tool files
Break down monolithic mcp-server.ts (3000+ lines) into modular structure:

- src/lib/mcp/types.ts - Shared types and interfaces
- src/lib/mcp/helpers.ts - Response formatting helpers
- src/lib/mcp/tools/*.ts - Tool registrars by domain:
  - ticket.ts (19 tools)
  - project.ts (9 tools)
  - board.ts (6 tools)
  - spec.ts (8 tools)
  - epic.ts (8 tools)
  - work.ts (5 tools)
  - workflow.ts (4 tools)
  - status.ts (5 tools)
  - phase.ts (6 tools)
  - action.ts (4 tools)
  - roadmap.ts (6 tools)
  - category.ts (4 tools)
  - template.ts (5 tools)
  - view.ts (3 tools)
  - cli-passthrough.ts (27 tools)

Main command file reduced from 3059 to 77 lines.
All 113 E2E tests pass.
