+++ apps/cli/src/lib/mcp/index.ts
@@ -1,0 +1,7 @@
+/**
+ * MCP Library - Main exports
+ */
+
+export * from './types.js'
+export * from './helpers.js'
+export * from './tools/index.js'
